Keep div_by_zero seperate
